requirements.txt
django==5.2.8
djangorestframework==3.16.1
requests==2.32.5
drf-yasg==1.21.11
python-dotenv==1.2.1